+++ impl/src/main/scala/akka/wamp/client/Publisher.scala
@@ -2,12 +2,14 @@ package akka.wamp.client
 
 import akka.Done
 import akka.actor.Actor._
+import akka.actor.Cancellable
 import akka.wamp._
 import akka.wamp.messages._
 import akka.wamp.serialization._
 
 import scala.collection.mutable
-import scala.concurrent.{Future, Promise}
+import scala.concurrent.duration._
+import scala.concurrent._
 
 /**
   * Publisher is a client which publishes events to topics
@@ -15,6 +17,9 @@ import scala.concurrent.{Future, Promise}
   */
 trait Publisher { this: Session =>
 
+  /** The Akka ActorSystem scheduler */
+  val scheduler = system.scheduler
+  
   /** The event publications map */
   private val pendingPublications: mutable.Map[PublicationId, Promise[Either[Done, Publication]]] = mutable.Map()
 
@@ -76,6 +81,29 @@ trait Publisher { this: Session =>
     */
   def publish(topic: Uri, ack: Boolean, data: List[Any], kwdata: Map[String, Any]): Future[Either[Done, Publication]] =  {
     publish(topic, ack, Payload(data, kwdata))
+  }
+
+  /**
+    * Publish a tick count to a topic with an initial delay and
+    * frequency. 
+    * 
+    * E.g. if you would like a message to be sent immediately and
+    * thereafter every 500ms you would set delay=Duration.Zero and
+    * interval=Duration(500, TimeUnit.MILLISECONDS)
+    * 
+    * @param topic is the topic to publish the tick count to
+    * @param delay is the initial delay 
+    * @param interval is the interval between two ticks
+    * @return
+    */
+  def tickle(topic: Uri, interval: FiniteDuration, delay: FiniteDuration = Duration.Zero): Future[Cancellable] = {
+    var count = 0
+    Future.successful(this.system.scheduler.schedule(delay, interval, new Runnable {
+      override def run(): Unit = {
+        count = count + 1
+        publish(topic, ack = false, List(count))
+      }
+    }))
   }